resource.h: Resource_Acquire takes u16, not u32 (dll_0b 98.60 -> 98.81)
include/main/resource.h declared Resource_Acquire(u32 id, int), contradicted by
dll_b7.c's file-local `extern void* Resource_Acquire(u16 id, int)`. The shadow is
the one that is RIGHT, so this inverts the usual "delete the shadow" reflex.

Proof is at dll_b7's call site: actionId is a u16, and retail passes it with
`mr r3,r30`. Under the u32 prototype MWCC must widen the u16 to u32 and emits
`clrlwi r3,r30,16` instead - dll_b7 only matches today *because* its local extern
overrides the header. Deleting the shadow against the u32 header changes codegen.

The callee is indifferent, which is why the wrong prototype survived: a u16
parameter is not guaranteed zero-extended by the ABI, so the callee zero-extends
on entry anyway and `clrlwi r0,r3,16` is emitted under both spellings. That also
makes the body's `index = id & 0xffff` dead under u16 (byte-identical without it),
so it reduces to `index = id` - the mask was only ever compensating for the u32.

Corroboration from an untouched unit: dll_0b calls Resource_Acquire((u16)(x+0x58), 1)
through the header. The extra u16->u32 conversion node the wrong prototype forced
was perturbing the register web; dropping it rotates dll_0B_func05's allocation
toward retail, 98.6034 -> 98.8071 (+0.20) with no source change in that file. A
wrong signature moving an unrelated unit *closer* to retail is the tell that u16
is the original.

With the header corrected, dll_b7's Resource_Acquire and Resource_Release shadows
are both redundant and removed in favour of the include (dll_b7.o byte-identical).

Gated in an origin clean-room on 6e9e1719a4: main.dol: OK, complete_units 800 flat,
matched_code FLAT, matched_data FLAT, fuzzy 99.84150 -> 99.84256, exactly one object
changed tree-wide (dll_0b, improved), zero regressions.

+++ include/main/resource.h
@@ -49,7 +49,7 @@ extern void* gResourceLoadedHandles[];
 extern u16 gResourceRefCounts[];
 
 BOOL Resource_Release(void *handleSlot);
-void *Resource_Acquire(u32 id, int unused);
+void *Resource_Acquire(u16 id, int unused);
 void Resource_ResetRefCounts(void);
 
 #endif /* MAIN_RESOURCE_H_ */
